scripts/previews_resized/requirements.txt
Pillow
pqdm
